Fixed a bug where changing the size of a textarea would make the page go crazy.
Since we were floating the textarea directly, once the size changed,
the result was undefined. CSS floating box model will try to use any
available space which is on the page, unless the size is fixed.
When the user is changing the size of a textarea the browser sets
the size to 100%, which would result in the undefined behaviour.